% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/format.float.R
\name{format.float}
\alias{format.float}
\title{Format numbers for printing}
\usage{
\method{format}{float}(x, digits = 2)
}
\arguments{
\item{x}{Value to be formatted.
Input can be a numeric or character value.}

\item{digits}{Number of fixed digits, after decimal point}
}
\value{
Returns a character value.
}
\description{
\code{format.float} prepares numbers for results printing.
}
\details{
This function formats numbers with the given number of decimal digits.
  Numbers are always assumed to be \code{\link[base]{double}}.
  If there is a percent sign in \code{x}, it is removed prior to formatting.
}
\examples{
format.float(0.5)
format.float(0.672, 4)
format.float(1.4927, 1)
format.float(c(1.756, 1.823))
# Input is character
format.float("1.12543")
format.float(c("1.756", "1.823"))
# Input is percentage
format.float("10\%", 1)
format.float("15 \%", 1)

# Useful for data frames
DF <- data.frame(num = c(1.756, 1.823), char = c("1.756", "1.823"), stringsAsFactors = FALSE)
transform(DF, num = format.float(num), char = format.float(char))

}
\author{
Felipe Figueiredo
}
