% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/format.pct.R
\name{format.pct}
\alias{format.pct}
\title{Convenience function to format proportions as percentages}
\usage{
\method{format}{pct}(p, digits = 1, pct.symbol = FALSE)
}
\arguments{
\item{p}{a numeric vector assumed to be a proportion.}

\item{digits}{the number of decimal digits to be used in the output.}

\item{pct.symbol}{(logical) whether or not to include the percent sign in the output.}
}
\value{
A character value formatted as percentages
}
\description{

}
\examples{
format.pct(.1)
format.pct(c(.42, .99), 0)

m <- 10*matrix(c(1, 1.1, .9, 2), nrow = 2)
format.pct(fisher.test(m)$p.value)
format.pct(fisher.test(5*m)$p.value)
format.pct(fisher.test(m)$conf.int)

}
\author{
Felipe Figueiredo
}
