% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/geocv.R
\name{geocv}
\alias{geocv}
\title{Compute the CV of log-normally distributed data.}
\usage{
geocv(x, na.rm = TRUE)
}
\arguments{
\item{x}{A numeric value, vector or list.}

\item{na.rm}{(logical) whether or not to remove \code{NA} values from computation}
}
\value{
Returns the geometric CV (as percentage)
}
\description{
\code{geocv} computes the CV of log-normal data.
}
\details{
This function first log10-transforms the original data \code{x}.
It then passes the standard deviation of the result to \code{logsd2cv}.
}
\examples{
d <- c(10,100,1000)
geocv(d)
geocv(d) == logsd2cv(1)

}
\author{
Felipe Figueiredo
}
