% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/geomean.R
\name{geomean}
\alias{geomean}
\title{Compute the geometric mean}
\usage{
geomean(x)
}
\arguments{
\item{x}{A numeric value, vector or list.}
}
\value{
Returns the geometric mean of \code{x}.
}
\description{
\code{geomean} computes the anti-log of the \code{mean} of log10-transformed data.
}
\details{
This function first log10-transforms the original data \code{x}.
It then back-transforms the \code{mean} of the results from the log10 scale.

If any null values are present in \code{x} they are first replaced with \code{NA}.
}
\examples{
d <- c(10,100,1000)
geomean(d)
geomean(c(d,0)) # na.rm is TRUE by default for sd()

}
\author{
Felipe Figueiredo
}
