% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/geosd.R
\name{geosd}
\alias{geosd}
\title{Compute the "geometric standard deviation" of data}
\usage{
geosd(x)
}
\arguments{
\item{x}{A numeric value, vector or list.}
}
\value{
The anti-log of the \code{sd} of log-transformed data `x`.
  Base 10 is assumed as a fixed default in this version.
}
\description{
The "geometric standard deviation" is not as well-defined as the geometric mean.
 This function returns the analogous concept of \code{sd} in the same manner of the geometric mean.
}
\examples{
d <- c(10,100,1000)
geosd(d)

}
\author{
Felipe Figueiredo
}
