% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/predint.R
\name{predint}
\alias{predint}
\title{Prediction Intervals for given data}
\usage{
predint(datavector, alpha = 0.05)
}
\arguments{
\item{datavector}{A numeric value, vector or list.}

\item{alpha}{the significance level required}
}
\value{
a vector of two values
  A vector with values giving lower and upper prediction limits for each parameter.
}
\description{
Compute a prediction interval of a vector, which is assumed
 normally distributed with unknown mean and variance.
}
\examples{
x <- rnorm(30)
predint(x)

}
\author{
Felipe Figueiredo
}
