% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/theme_ff_table.R
\name{theme_ff_gtsummary}
\alias{theme_ff_gtsummary}
\title{theme_ff_gtsummary}
\usage{
theme_ff_gtsummary(set_theme = TRUE, font_size = NULL)
}
\arguments{
\item{set_theme}{Set the FF gtsummary theme, Default: TRUE}

\item{font_size}{Numeric font size for compact theme. Default is 13 for gt tables, and 8 for all other output types, Default: NULL}
}
\value{
OUTPUT_DESCRIPTION
}
\description{
FF gtsummary theme
}
\details{
DETAILS
}
\examples{
theme_ff_gtsummary()
}
\seealso{
\code{\link[gtsummary]{set_gtsummary_theme}}
}
