% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cv2logvar.R
\name{cv2logvar}
\alias{cv2logvar}
\title{Convert the CV from a log-normal variable to the log-transformed variance.}
\usage{
cv2logvar(x)
}
\arguments{
\item{x}{A numeric value, vector or list.}
}
\value{
Returns the log-transformed variance.
}
\description{

}
\examples{
cv2logvar(50)
# [1] 0.04208748

}
\author{
Felipe Figueiredo
}
