% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/format.interval.R
\name{format.interval}
\alias{format.interval}
\title{Format intervals for printing}
\usage{
\method{format}{interval}(interval, digits = 2)
}
\arguments{
\item{interval}{Must be a vector of two values.
Values can be either numeric or character.}

\item{digits}{Number of fixed digits, after decimal point.}
}
\value{
Returns a character value.
}
\description{
\code{format.interval} prepares intervals for results printing.
}
\details{
This function formats a given interval with the given number of decimal digits.
}
\examples{
format.interval(c(1.756, 1.823))
# format.interval(1.5, 1) # error: 1.5 is not an interval

# Input is character
format.interval(c("1.756", "1.823"))

}
\author{
Felipe Figueiredo
}
