% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/format.pct.R
\name{format.pct}
\alias{format.pct}
\title{Format proportions as percentages for printing}
\usage{
\method{format}{pct}(p, digits = 1, pct.symbol = TRUE)
}
\arguments{
\item{p}{a numeric vector assumed to be a proportion.}

\item{digits}{the number of decimal digits to be used in the output.}

\item{pct.symbol}{(logical) whether or not to include the percent sign in the output.}
}
\value{
A character value formatted as percentages
}
\description{
\code{format.pct} prepares proportions as percentages for results printing, possibly with a percent sign.
}
\details{
This function multiplies the proportion by 100, formats with the given number of decimal digits.
It also (optionally) includes a percent sign.
}
\examples{
format.pct(.1)
format.pct(c(.42, .99), 0)

m <- 10*matrix(c(1, 1.1, .9, 2), nrow = 2)
format.pct(fisher.test(m)$p.value)
format.pct(fisher.test(5*m)$p.value)
format.pct(fisher.test(m)$conf.int)

}
\author{
Felipe Figueiredo
}
