% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/logsd2cv.R
\name{logsd2cv}
\alias{logsd2cv}
\title{Convert the log-transformed standard deviation from a log-normal variable to the CV.}
\usage{
logsd2cv(x)
}
\arguments{
\item{x}{A numeric value, vector or list.}
}
\value{
Returns the geometric CV (as percentage)
}
\description{

}
\examples{
d <- c(10,100,1000)
sd(log10(d)) # 1
logsd2cv(1) # 1413.21\\\%

}
\author{
Felipe Figueiredo
}
