% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/logvar2cv.R
\name{logvar2cv}
\alias{logvar2cv}
\title{Convert the log-transformed variance from a log-normal variable to the CV.}
\usage{
logvar2cv(x)
}
\arguments{
\item{x}{A numeric value, vector or list.}
}
\value{
Returns the geometric CV (as percentage)
}
\description{

}
\examples{
d <- c(10,100,1000)
var(log10(d)) # 1
logvar2cv(1) # 1413.21\\\%

}
\author{
Felipe Figueiredo
}
